logging/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
grpcio==1.70.0
grpcio-tools==1.70.0
hazelcast-python-client==5.5.0
idna==3.10
protobuf==5.29.3
py-consul==1.6.0
requests==2.32.3
setuptools==75.8.2
urllib3==2.3.0
